feat: add service type

+++ dandelion/crud/base.py
@@ -73,8 +73,9 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
     def remove(self, db: Session, *, id: int) -> ModelType:
         obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
+        if obj:  # TODO(Wu.Wenxiang) query.delete or select
+            db.delete(obj)
+            db.commit()
         return obj
 
     @staticmethod
